+++ src/Generic/FourOhFour.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Link} from "react-router-dom";
 
-export default function FourOhFour(props){
+export default function FourOhFour(){
     return (
         <div className="center fourOhFour">
             <h3>Wygląda na to, iż próbowałeś wejść na stronę, która już nie istnieje lub wszedłeś tutaj przypadkowo.</h3>